+++ ui/expenselist/src/main/java/com/kappzzang/jeongsan/expenselist/MemberAdapter.kt
@@ -1,4 +1,4 @@
-package com.kappzzang.jeongsan.creategroup
+package com.kappzzang.jeongsan.expenselist
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,7 +6,6 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.kappzzang.jeongsan.data.Member
-import com.kappzzang.jeongsan.expenselist.R
 
 class MemberAdapter(var memberList: List<Member>, val inflater: LayoutInflater, val layoutId: Int) :
     RecyclerView.Adapter<MemberAdapter.MemberViewHolder>() {
